refactor(messages): separate command responsibilities and date parsing
